fix(shortcuts): Shortcut urls are case sensitive
Fixes an issue where the shortcut urls are case sensitive, which would
cause any case-insensitive URL requests to not result in a 404 when the
case does not match up.

Modified the Shortcuts' shortcutByCode server controller method to use
an aggregate query to force a lowercase comparison to find the matching
Shortcut.

Fixes #1